+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.4.0] 2025-02-03
 ### Added
 - Screenshots: `screenshot/{2,3}` function and `screenshot: true` config for auto-capture
+- Pass browser config via `Case` opts, e.g. `use PhoenixTest.Playwright.Case, headless: false`
 
 ### Changed
 - Renamed: `PheonixTest.Case` to `PhoenixTest.Playwright.Case`